+++ backend/Application/Api/GraphQL/Blocks/BalanceStatistics.cs
@@ -4,12 +4,22 @@ namespace Application.Api.GraphQL.Blocks;
 
 public class BalanceStatistics
 {
-    public BalanceStatistics(ulong totalAmount, ulong? totalAmountReleased, ulong totalAmountEncrypted, 
-        ulong totalAmountLockedInReleaseSchedules, ulong totalAmountStaked, ulong totalAmountStakedByBakers, 
-        ulong totalAmountStakedByDelegation, ulong bakingRewardAccount, ulong finalizationRewardAccount, ulong gasAccount)
+    public BalanceStatistics(
+        ulong totalAmount, 
+        ulong? totalAmountReleased, 
+        ulong? totalAmountUnlocked, 
+        ulong totalAmountEncrypted, 
+        ulong totalAmountLockedInReleaseSchedules, 
+        ulong totalAmountStaked, 
+        ulong totalAmountStakedByBakers, 
+        ulong totalAmountStakedByDelegation, 
+        ulong bakingRewardAccount, 
+        ulong finalizationRewardAccount, 
+        ulong gasAccount)
     {
         TotalAmount = totalAmount;
         TotalAmountReleased = totalAmountReleased;
+        TotalAmountUnlocked = totalAmountUnlocked;
         TotalAmountEncrypted = totalAmountEncrypted;
         TotalAmountLockedInReleaseSchedules = totalAmountLockedInReleaseSchedules;
         TotalAmountStaked = totalAmountStaked;
@@ -23,8 +33,11 @@ public class BalanceStatistics
     [GraphQLDescription("The total CCD in existence")]
     public ulong TotalAmount { get; init; }
     
-    [GraphQLDescription("The total CCD released according to the Concordium promise published on deck.concordium.com. Will be null for blocks with slot time before the published release schedule.")]
+    [GraphQLDescription("The total CCD Released. This is total CCD supply not counting the balances of non circulating accounts")]
     public ulong? TotalAmountReleased { get; init; }
+
+    [GraphQLDescription("The total CCD Unlocked according to the Concordium promise published on deck.concordium.com. Will be null for blocks with slot time before the published release schedule.")]
+    public ulong? TotalAmountUnlocked { get; init; }
 
     [GraphQLDescription("The total CCD in encrypted balances")]
     public ulong TotalAmountEncrypted { get; init; }