+++ backend/CHANGELOG.md
@@ -11,4 +11,6 @@
 - Added boolean filters on Bakers Query to enable filtering on Baker Removed State.
 - Fixed Memo CBOR Decoding.
 - Adds Listing of Contracts Via View
+- Changed the header from "Total circulating supply of CCD" to "Total unlocked CCD"
+- Added New API endpoint to get "unlocked CCD"
 - Fixed Issue 55 by setting account balance to actual queried balance in case of genesis block.